refactor boundless code into own module
