+++ bin/cli/Cargo.toml
@@ -34,8 +34,6 @@ kona-host.workspace = true
 kona-derive = { workspace = true }
 
 bonsai-sdk.workspace = true
-boundless-market.workspace = true
-risc0-aggregation.workspace = true
 risc0-ethereum-contracts.workspace = true
 risc0-zkvm.workspace = true
