+++ bin/client/Cargo.toml
@@ -32,6 +32,7 @@ kona-std-fpvm.workspace = true
 
 bonsai-sdk.workspace = true
 boundless-market.workspace = true
+risc0-aggregation.workspace = true
 risc0-ethereum-contracts.workspace = true
 risc0-zkvm.workspace = true
