+++ book/src/setup.md
@@ -43,7 +43,7 @@ kailua-cli config --op-node-url [YOUR_OP_NODE_URL] --op-geth-url [YOUR_OP_GETH_U
 Running the above command against the respective op-sepolia endpoints should produce the following output:
 ```
 RISC0_VERSION: 1.2.0
-FPVM_IMAGE_ID: 0x4A95CC2ED74806395AD1E8AA10A7843261FFB6333D1DC45BBB1FF965490288AD
+FPVM_IMAGE_ID: 0x450100AAB9270940ECCE2D7F4276767E037018A522BBC99F1089D52E494AFF28
 CONTROL_ROOT: 0x8CDAD9242664BE3112ABA377C5425A4DF735EB1C6966472B561D2855932C0469
 CONTROL_ID: 0x04446E66D300EB7FB45C9726BB53C793DDA407A62E9601618BB43C5C14657AC0
 SET_BUILDER_ID: 0x744CCA56CDE6933DEA72752C78B4A6CA894ED620E8AF6437AB05FAD53BCEC40A
